bmi A function to read all available tabs from an excel file

bmi A function to read all available tabs from an excel file

multi_import_excel(file = "inst/corrupted_toydatasets.xlsx")

Arguments

file

the name or pathfile of the excel file (type = character)

Value

A list of data frames named after the excel tabs

Examples

file <- system.file("corrupted_toydatasets.xlsx", package = "annamantsoki")
multi_import_excel(file)
#> $diamonds
#> $diamonds$result
#> # A tibble: 50 × 10
#>    carat cut       color clarity depth table price     x     y     z
#>    <dbl> <chr>     <chr> <chr>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  1.21 Premium   D     VS1      60.2  59   10083  6.89  6.86  4.14
#>  2  0.54 Ideal     D     VS1      62.1  57    2004  5.19  5.22  3.23
#>  3  0.31 Very Good I     SI1      63.5  56     523  4.32  4.28  2.73
#>  4  0.7  Ideal     F     VS1      62.5  56    2751  5.73  5.7   3.57
#>  5  1.2  Ideal     F     VVS1     61.1  55   13088  6.86  6.89  4.2 
#>  6  0.56 Ideal     H     SI2      60.7  56    1571  5.34  5.37  3.25
#>  7  1.21 Fair      G     SI1      64.6  60    6397  6.69  6.62  4.3 
#>  8  0.53 Ideal     H     VVS1     61.3  56    1781  5.22  5.24  3.21
#>  9  0.72 Premium   H     VS2      61.6  57    2418  5.84  5.78  3.58
#> 10  0.61 Very Good F     VVS2     60.5  58.5  2274  5.44  5.64  3.34
#> # … with 40 more rows
#> 
#> $diamonds$error
#> NULL
#> 
#> 
#> $cars
#> $cars$result
#> NULL
#> 
#> $cars$error
#> <simpleError: Sheet 'cars' (position 2): Invalid sheet xml (no <sheetData>)>
#> 
#> 
#> $iris
#> $iris$result
#> # A tibble: 25 × 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species   
#>           <dbl>       <dbl>        <dbl>       <dbl> <chr>     
#>  1          4.9         3            1.4         0.2 setosa    
#>  2          5.7         4.4          1.5         0.4 setosa    
#>  3          7.2         3.6          6.1         2.5 virginica 
#>  4          6.2         2.8          4.8         1.8 virginica 
#>  5          6.3         3.4          5.6         2.4 virginica 
#>  6          6.8         3.2          5.9         2.3 virginica 
#>  7          7.7         3            6.1         2.3 virginica 
#>  8          5.1         3.7          1.5         0.4 setosa    
#>  9          4.9         3.1          1.5         0.2 setosa    
#> 10          5.6         2.5          3.9         1.1 versicolor
#> # … with 15 more rows
#> 
#> $iris$error
#> NULL
#> 
#>